Split dom styles into smaller modules
