+++ src/dom/tests/createElement.test.ts
@@ -1,0 +1,4 @@
+import { beforeAll, describe, it, expect } from "vitest";
+import { addClass, getIds, IterableElements, removeClass, style } from "..";
+
+describe("createElement", () => {})
